+++ src/eventHandlers/guildMemberHandler.ts
@@ -3,7 +3,7 @@ import client from "../util/client";
 import Channels from "../util/channels";
 import { mentionChannelWithId } from "../helpers";
 import isDevEnv from "../util/isDevEnv";
-import { ranksRepo, inventoryRepo } from "../util/collections_v2";
+import { ranksRepo, inventoryRepo } from "../util/collections";
 
 export const guildMemberAddHandler = async (member: GuildMember) => {
     const greeting = `**Welcome to Yuqi's Cookie House :cookie: ${member.toString()}!**\nRules and other information is available in ${mentionChannelWithId(Channels.Reception.INFO)}.\nGrab your roles from ${mentionChannelWithId(Channels.Reception.ROLES)} and ask staff if you need anything!`;
