+++ src/jobs/index.ts
@@ -1,3 +1,4 @@
 export * from "./updateEmotes";
 export * from "./triggerEvent";
 export * from "./help";
+export * from "./tempMigrateRankToProfile";
